requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
websockets==12.0

# Cryptography
pynacl==1.5.0
eth-account==0.10.0
web3==6.11.3
cryptography==41.0.7

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Async Support
aiohttp==3.9.1

# Monitoring & Logging
rich==13.7.0
prometheus-client==0.19.0

# Web Dashboard
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
